+++ tests/SmartRide.UnitTests/Application/Handlers/Vehicles/GetVehicleByVinQueryHandlerTests.cs
@@ -55,7 +55,7 @@ public class GetVehicleByVinQueryHandlerTests
             RegisteredDate = vehicle.RegisteredDate,
             VehicleType = new VehicleTypeDTO
             {
-                VehicleTypeId = (byte)vehicle.VehicleTypeId,
+                VehicleTypeId = vehicle.VehicleTypeId,
                 Name = vehicle.VehicleTypeId.ToString()
             },
         };
